Rename ConfigStore
