+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
-import store from './redux/configureStore';
+import store from './redux/ConfigStore';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
